Refactor y ajustes
