+++ .dockerignore
@@ -1,3 +1,4 @@
 venv
 postgres-data
 .env
+*.log
